+++ Changelog.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New `toolbox` command group.
 - `toolbox check-dep` command to validate single `dep` files.
 - `vivado check-syntax` command to check vhdl syntax before synthesis.
-- Colorised Vivado messages.
+- Colorization of Vivado messages.
 
 ## [0.3.0] - 2018-03-11
 ### Changed