+++ Changelog.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-##[v0.3.2] - 2018-5-25
+# [Unreleased]
+### Changed
+-
+
+### Added
+-
+
+## [0.3.2] - 2018-5-25
 ### Changed
 - Improved `srcs status` summary for git tags.
 - Modelsim workflow updated. Xilinx simlib generation moved to a dedicated command `setup-simlib`.